requirements.txt
Flask
google-cloud-datastore
gunicorn
flask-cors
